improved write_*, fixes and master rebase

+++ src/alert.cpp
@@ -1981,7 +1981,7 @@ namespace libtorrent {
 			tcp::endpoint const& endp = peers[i];
 			std::size_t const size = endp.size();
 			TORRENT_ASSERT(size < 0x100);
-			detail::write_uint8(uint8_t(size), ptr);
+			detail::write_uint8(size, ptr);
 			std::memcpy(ptr, endp.data(), size);
 			ptr += size;
 		}
